[RF] Re-implement `MakeNumpyDataFrame` with no interpreter calls
Re-implement `MakeNumpyDataFrame` with no interpreter calls, and avoid a
explicit lifetime management of Python objects in the RDNumpyDS helper.

Like this, the former `RNumpyDS.hxx` helper could be moved to
`tree/dataframe`, as it doesn't import `Python.h` anymore.

The `RNumpyDS` also got renamed to `RVecDS`, because it takes RVecs and
not NumPy arrays directly.