[CMake] Fix ordering problem when checking and setting `cuda` flag
In the ROOT build system, `CheckCompiler` is included before
`RootBuildOptions`. The problem with this is that when building with
`-Dall=ON`, the `cuda` flag is only set in `RootBuildOptions`, which is
too late to find the CUDA compiler in `CheckCompiler`. This results in
issues like #15139:
```
CMake Error in roofit/batchcompute/CMakeLists.txt:
  Exporting the target "RooBatchCompute_CUDA" is not allowed since its linker
  language cannot be determined
```

This commit suggests a solution to this problem: just always enable the
CUDA language in CMake if it is found, and only set the `cuda` flag with
`-Dall=ON` if the CUDA language is enabled.