[TO REVERT] Never run TMVA GNN tests and tutorials as unit tests
The last commit adds the dependencies for the TMVA GNN unit tests to the
docker images via the `requirements.txt`. However, this will only have a
delayed effect until the images are re-built. Therefore, we can't
validate for now that the tests actually work.

Once the missing packages make it into the CI images, a PR should be
opened to revert this commit.