[RF] error message when fit ends with only bad NLL evals
The RooNaNPacker-based error handling introduced a subtle uncaught "failing" fit that reports as a successful fit. The issue arises when immediately from the first NLL call, there are no valid NLL values being returned. This means that _maxFCN is still the default value -inf, which means that in RooAbsMinimizerFcn::applyEvalErrorHandling, where the fvalue is set to `(std::isfinite(_maxFCN) ? _maxFCN : 0.) + cfg().recoverFromNaN * badness`, if the badness is also zero (i.e. when the value is not a packed NaN; it may for instance be a regular NaN or some other evaluation error may have occurred), the fvalue will be zero. If the next evaluation again gives errors, the value will again be zero, etcetera. Minuit will then think that it is a very nicely converged function with value zero. It will thus report a successful fit.

This commit adds a RooFit error-level message. Perhaps the response should be stronger (change the status number, do something inside the RooFitResult as well, throw an exception, etc.), but an error message seems like the bare minimum courtesy to the user.

Note that the LikelihoodGradientJobBinnedErrorTest.TriggerMuLEZero test
triggers this error already.