[ntuple] Create processor fields given an `RNTupleModel`
The underlying entry used by the processor is build based on the
provided model. The value pointers in the model's default entry are used
to hold the field values during iteration. This way, the pointers
returned by `RNTupleModel::MakeField` can be used in the processor
iteration to access the values being read.