[core] Correct the regular expression for the scheme part in TUri class
RFC 3986 (https://datatracker.ietf.org/doc/html/rfc3986) defines the
scheme part of a URI as:

    scheme      = ALPHA *( ALPHA / DIGIT / "+" / "-" / "." )

The regular expression for finding the scheme part in the TUri class
was defined as

    ^[[:alpha:]][[:alpha:][:digit:]+-.]*$

This does not match the definition in the RFC, since +-. in regular
expression syntax is the range of ascii codes from '+' to '.', which
means '+', ',', '-' and '.'. I.e. the ',' is included in the regular
expression in error. This commit adds a backslash escape to the '-' so
that it is interpreted as a literal '-' sign instead of indicating a
range in the regular expression.