[ntuple] remove unused column flags
The flags for sorted and non-negative columns are currently unused.
Moreover, they don't carry any more information than is provide by the
column type (and, in its current implementation, they where wrong in
addition).

Such aggregate information about the values in a column may be useful in
the future but only if it takes into account the actual values written.
In this case, the header is the wrong place to store the information.