[PyROOT] Call typed overload of `SetBranchAddress` for array types
Call typed overload of `SetBranchAddress` for array types like
`np.ndarray` or `array.array`. The correct C++ template parameter is
inferred from the array interface type information.

This is an improvement for two reasons:

  1. We don't need to cast the arrays to a `void *` to avoid template
     resolution problems anymore.

  2. More importantly: ROOT will to a runtime check for type
     compatibility, which was requested in the JIRA ticked linked below.

This change is also mentioned in the release notes and tested by a new
unit test.

To see the benefit of the added type check, take for example this code,
where the branch address is set to an array address with a mismatched
type:

```python
import array
import ROOT

def create():
    f = ROOT.TFile("dummy.root", 'RECREATE')
    t = ROOT.TTree("tree", 'A tree')
    a = array.array('f', [42])
    br = t.Branch('a', a, 'a/F')

    for x in (21, 42, 1337):
        a[0] = x
        t.Fill()

    t.Write()
    f.Close()

def test(typecode='f'):
    f = ROOT.TFile("dummy.root", 'READ')
    t = f.Get('tree')
    a = array.array(typecode, [0])
    status = t.SetBranchAddress('a', a)
    print("Status = %s" % (status, ))

    for i in range(t.GetEntries()):
        t.GetEntry(i)
        print("a = %s" % (a[0], ))

create()
test("f")
test("d")
```

Before this commit, the output would be like this:
```txt
Status = 4
a = 21.0
a = 42.0
a = 1337.0
Status = 4
a = 5.442276803e-315
a = 5.483722033e-315
a = 5.690664868e-315
```

With this commit, you get a clear error:
```txt
Status = 0
a = 21.0
a = 42.0
a = 1337.0
Error in <TTree::SetBranchAddress>: The pointer type given "Double_t" (8) does not correspond to the type needed "Float_t" (5) by the branch: a
Status = -2
a = 0.0
a = 0.0
a = 0.0
```

Closes the following JIRA ticket:
https://its.cern.ch/jira/browse/ROOT-8213